Put Submission thunk actions in actions.thunk file
- moved from "Redux component" files
